feat: mejorar la gestión de limpieza al actualizar el estado de la habitación
